requirements.txt
requests
colorama
psutil
python-whois
dnspython
